sérieux début

+++ libft/lib_fct/lib_str/src/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: eslamber <eslamber@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: dbaule <dbaule@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 14:07:50 by eslamber          #+#    #+#             */
-/*   Updated: 2022/11/25 14:15:52 by eslamber         ###   ########.fr       */
+/*   Updated: 2023/10/29 17:03:45 by dbaule           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ static int	ft_iswhitespace(int car)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	long long	res;
+	long		res;
 	size_t		ind;
 	int			minus;
 
@@ -49,5 +49,5 @@ int	ft_atoi(const char *str)
 		ind++;
 	}
 	res = res * minus;
-	return ((int) res);
+	return (res);
 }
